Fix: Resolve NameError in Trainer.evaluate for Levenshtein task
I corrected a `NameError` that occurred in `Trainer.evaluate` within `train_loop.py` when `use_levenshtein_task` was enabled.

The errors were:
1. `NameError: name 'true_lev_distances' is not defined`
2. `NameError: name 'is_shuffled_flags' is not defined`

This commit addresses these issues as follows:
- The `true_lev_distances` required for calculating the Levenshtein auxiliary loss are now correctly derived by filtering the `auxiliary_values` tensor from the batch data using a mask where `task_type_flags == 1.0` (indicating Levenshtein task items).
- The predicted Levenshtein distances from the model output are similarly filtered using the same mask before computing the loss.
- The logic for monitoring predicted Levenshtein distances on 'original' items (i.e., LM task items) now correctly uses `task_type_flags == 0.0` to identify these items, replacing the undefined `is_shuffled_flags`.

These changes ensure that the evaluation logic for the Levenshtein auxiliary task correctly identifies and processes the relevant data from the batch, resolving the `NameError` and ensuring accurate loss calculation and metric monitoring.

+++ nsp_dataset_new.py
@@ -102,6 +102,6 @@ class NSPDataset(Dataset):
         return (
             torch.tensor(input_tokens, dtype=torch.long),
             torch.tensor(lm_targets, dtype=torch.long),
-            torch.tensor(nsp_class, dtype=torch.long),  # NSP class label
+            torch.tensor(float(nsp_class), dtype=torch.float),  # NSP class label
             torch.tensor(2.0, dtype=torch.float),  # Task type flag (2.0 for NSP)
         )
